Updated build_helper.sh for ruby 3.4.0 later

+++ .github/workflows/build_helper.sh
@@ -1277,13 +1277,16 @@ if [ "${CI_DO_PUBLISH}" -eq 1 ]; then
 			PRNERR "Failed to install addrev gem"
 			exit 1
 		fi
+	else
+		PRNINFO "Skip to install addrev gem"
+	fi
+
+	if [ "${IS_OS_FEDORA}" -eq 1 ] && echo "${CI_OSTYPE}" | sed -e 's#:##g' | grep -q -i 'fedora42'; then
 		PRNINFO "Start to install json gem"
-		if ({ RUNCMD "${GEM_BIN}" "${GEM_INSTALL_CMD}" json || echo > "${PIPEFAILURE_FILE}"; } | sed -e 's/^/    /g') && rm "${PIPEFAILURE_FILE}" >/dev/null 2>&1; then
+		if ({ RUNCMD "${GEM_BIN}" "${GEM_INSTALL_CMD}" json -v 2.9.1 || echo > "${PIPEFAILURE_FILE}"; } | sed -e 's/^/    /g') && rm "${PIPEFAILURE_FILE}" >/dev/null 2>&1; then
 			PRNERR "Failed to install json gem"
 			exit 1
 		fi
-	else
-		PRNINFO "Skip to install addrev and json gem"
 	fi
 else
 	PRNINFO "Skip to install published tools for uploading packages to packagecloud.io, because this CI process does not upload any packages."
